KnowledgeSelection/requirement.txt
charset-normalizer==2.1.0
click==8.1.3
filelock==3.7.1
idna==3.3
importlib-metadata==4.12.0
jieba==0.42.1
joblib==1.1.0
numpy==1.21.6
packaging==21.3
pyparsing==3.0.9
pytorch-crf==0.7.2
regex==2022.7.9
requests==2.28.1
sacremoses==0.0.53
six==1.16.0
tokenizers==0.10.3
torch==1.8.1
tqdm==4.64.0
transformers==4.5.1
typing_extensions==4.3.0
urllib3==1.26.10
zipp==3.8.1
